backend/requirements.txt
# SwarmCourt Backend Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Solana interaction
solders>=0.21.0
solana>=0.34.0
anchorpy>=0.20.1
base58>=2.1.0

# LLMs for real debates
groq>=0.5.0
openai>=1.0.0
google-genai>=0.1.0

# IPFS storage
requests>=2.31.0
aiohttp>=3.9.0
